# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Foundry - Decentralized AI Model Fine-Tuning Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List all marketplace models
      parameters:
        - name: baseModel
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [popular, newest, cheapest]
        - name: creatorWallet
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
  /models/{id}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get model details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Model details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Not found
    patch:
      operationId: updateModel
      tags: [models]
      summary: Update model listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelBody"
      responses:
        "200":
          description: Updated model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
  /models/{id}/list:
    post:
      operationId: listModel
      tags: [models]
      summary: List a model on the marketplace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListModelBody"
      responses:
        "200":
          description: Model listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
...
components:
  schemas:
...
    Model:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        nftTokenId:
          type: ["string", "null"]
        ogChainTxHash:
          type: ["string", "null"]
        ogExplorerUrl:
          type: ["string", "null"]
        creatorWallet:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        baseModel:
          type: string
        datasetDescription:
          type: string
        samplePrompt:
          type: string
        sampleOutput:
          type: string
        licensePriceUsd:
          type: number
        isListed:
          type: boolean
        inferenceCount:
          type: integer
        licenseCount:
          type: integer
        modelRootHash:
          type: ["string", "null"]
        datasetRootHash:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - creatorWallet
        - name
        - description
        - category
        - baseModel
        - datasetDescription
        - samplePrompt
        - sampleOutput
        - licensePriceUsd
        - isListed
        - inferenceCount
        - licenseCount
        - createdAt
    UpdateModelBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        licensePriceUsd:
          type: number
        isListed:
          type: boolean
        samplePrompt:
          type: string
        sampleOutput:
          type: string
    ListModelBody:
      type: object
      properties:
        licensePriceUsd:
          type: number
        creatorWallet:
          type: string
      required:
        - licensePriceUsd
        - creatorWallet
...