# lib/api-zod/src/generated/types/updateModelBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Foundry - Decentralized AI Model Fine-Tuning Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateModelBody {
  name?: string;
  description?: string;
  licensePriceUsd?: number;
  isListed?: boolean;
  samplePrompt?: string;
  sampleOutput?: string;
}